- Refactor renamed package "legacy" to "versioning" for better code practice. - Changed naming convention to ...CompatibilityManager for better code practice.
