+++ src/main/java/dev/thesheep/simpleresourcepack/api/players/ResourcepackEvents.java
@@ -1,8 +1,8 @@
 package dev.thesheep.simpleresourcepack.api.players;
 
 import dev.thesheep.simpleresourcepack.SimpleResourcepack;
-import dev.thesheep.simpleresourcepack.legacy.ActionBar;
-import dev.thesheep.simpleresourcepack.legacy.ResourcePack;
+import dev.thesheep.simpleresourcepack.versioning.ActionBarCompatibilityManager;
+import dev.thesheep.simpleresourcepack.versioning.ResourcePackCompatibilityManager;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -24,11 +24,11 @@ public class ResourcepackEvents implements Listener {
         if(event.getStatus() == PlayerResourcePackStatusEvent.Status.SUCCESSFULLY_LOADED)
         {
             String msg = SimpleResourcepack.getInstance().getConfig().getString("message_download_complete", "");
-            ActionBar.sendActionBar(event.getPlayer(), msg);
+            ActionBarCompatibilityManager.sendActionBar(event.getPlayer(), msg);
         }
-        else if (ResourcePack.isUsingLegacyMethod()
+        else if (ResourcePackCompatibilityManager.isUsingLegacyMethod()
                 && event.getStatus() == PlayerResourcePackStatusEvent.Status.DECLINED
-                && ResourcePack.isForced(event.getPlayer()))
+                && ResourcePackCompatibilityManager.isForced(event.getPlayer()))
         {
             // TODO: Consider making this message configurable in config.yml (See if it's also possible to do the same for 1.20+ for consistency)
             event.getPlayer().kickPlayer("You must accept the resource pack to play on this server.\nEnsure you set Edit Server > Server Resource Packs to Enabled");
@@ -36,7 +36,7 @@ public class ResourcepackEvents implements Listener {
         else if(event.getStatus() == PlayerResourcePackStatusEvent.Status.FAILED_DOWNLOAD)
         {
             String msg = SimpleResourcepack.getInstance().getConfig().getString("message_download_failed", "");
-            ActionBar.sendActionBar(event.getPlayer(), msg);
+            ActionBarCompatibilityManager.sendActionBar(event.getPlayer(), msg);
         }
     }
